ENG-105 plugin list, refactor
