+++ src/ui/app/AppContainer.js
@@ -1,5 +1,4 @@
 import { connect } from 'react-redux';
-import { fetchPlugins } from 'state/plugins/thunks';
 import { withRouter } from 'react-router-dom';
 import { getUsername } from '@entando/apimanager';
 import App from 'ui/app/App';
@@ -9,10 +8,6 @@ export const mapStateToProps = (state, { location: { pathname } }) => ({
   currentRoute: pathname,
 });
 
-export const mapDispatchToProps = dispatch => ({
-  fetchPlugins: () => dispatch(fetchPlugins()),
-});
-
-const AppContainer = withRouter(connect(mapStateToProps, mapDispatchToProps)(App));
+const AppContainer = withRouter(connect(mapStateToProps, null)(App));
 
 export default AppContainer;
