+++ src/ui/plugins/PluginListPageContainer.js
@@ -2,10 +2,16 @@ import { connect } from 'react-redux';
 import { getPluginList } from 'state/plugins/selectors';
 import PluginListPage from 'ui/plugins/PluginListPage';
 import { getLoading } from 'state/loading/selectors';
+import { fetchPlugins } from 'state/plugins/thunks';
 
 export const mapStateToProps = state => ({
   plugins: getPluginList(state),
   loading: getLoading(state).plugins,
 });
 
-export default connect(mapStateToProps, null)(PluginListPage);
+
+export const mapDispatchToProps = dispatch => ({
+  fetchPlugins: () => dispatch(fetchPlugins()),
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(PluginListPage);
